pyproject.toml
[tool.poetry]
name = "fastapi_reclamos"
version = "0.1.0"
description = "RESTful API for reclamos app"
authors = ["dfcarrera79 <dfcarrera@outlook.com>"]
readme = "README.md"
packages = [{include = "fastapi_reclamos"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.11"
psycopg2 = "^2.9.6"
pillow = "^9.5.0"
python-multipart = "^0.0.6"
httpx = "^0.24.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
